requirements.txt
pandas==2.2
fastapi==0.111
pyarrow==15.0
openpyxl==3.0.10
uvicorn==0.18.3
mysql-connector-python~=8.0.32
pydantic~=1.10.5
starlette~=0.36
psycopg2~=2.9
boto3~=1.34
redis==5.0.3
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp==1.24.0
isort~=5.13
black~=24.4